// Backend-Horarius/src/modules/clients/dtos/client.dto.ts
export type ClientDto = {
  id: number;
  name: string;
  email: string;
  phone: string;
  cpf: string;
  notes: string;
};

export type CreateClientRequestDto = {
  name: string;
  email: string;
  phone: string;
  cpf?: string;
  notes: string;
};

export type UpdateClientRequestDto = {
  name: string;
  email: string;
  phone: string;
  cpf?: string;
  notes: string;
};

export type ListClientsQueryDto = {
  limit?: number;
  page?: number;
  search?: string;
};